README Fix: Remove ponit about MAC address issue
The claim that UUIDs require access to the MAC address should be removed, as it's not technically correct.  RFC4122 allows for the `node` field to be set to a random value if a MAC address isn't available. Ref: https://datatracker.ietf.org/doc/html/rfc4122#section-4.1.6

FWIW, the `uuid` module has always used a random value for the `node` field and never had any complaints.  In fact, this ultimately turned out to be a good thing, as embedding MAC addresses in IDs that might be shared publicly has come to be seen as something of a security concern.  I guess if you're looking for reasons to take down UUID you could list that instead?  But where `uuid` is concerned it's not an issue.